separated plotting tests in new file test_plotting.py (currently not run in Travis) to fix broken Travis checks
